@@ -4,27 +4,27 @@
   "dependencies": {
     "@hotosm/iso-countries-languages": "^1.1.2",
     "@karaokemugen/i18n-iso-languages": "^3.1.1",
-    "@sentry/integrations": "^6.19.2",
-    "@sentry/react": "^6.19.2",
-    "antd": "4.19.3",
+    "@sentry/integrations": "^6.19.3",
+    "@sentry/react": "^6.19.3",
+    "antd": "4.19.5",
     "dompurify": "^2.3.6",
     "i18next": "^21.6.14",
     "lodash": "^4.17.21",
     "minireset.css": "^0.0.7",
     "path-browserify": "^1.0.1",
     "pretty-bytes": "^6.0.0",
-    "react": "17.0.2",
+    "react": "18.0.0",
     "react-aria-menubutton": "^7.0.3",
     "react-beautiful-dnd": "^13.1.0",
-    "react-dom": "17.0.2",
-    "react-i18next": "^11.16.1",
-    "react-image-crop": "^9.1.1",
-    "react-router-dom": "6.2.2",
+    "react-dom": "18.0.0",
+    "react-i18next": "^11.16.2",
+    "react-image-crop": "^10.0.0",
+    "react-router-dom": "6.3.0",
     "react-text-transition": "^1.3.0",
     "react-toastify": "^8.2.0",
-    "react-virtuoso": "^2.8.3",
+    "react-virtuoso": "^2.8.6",
     "resize-observer-polyfill": "^1.5.1",
-    "sass": "^1.49.9",
+    "sass": "^1.49.11",
     "slugify": "^1.6.5",
     "socket.io-client": "4.4.1",
     "use-async-memo": "^1.2.3",
@@ -32,8 +32,8 @@
   },
   "devDependencies": {
     "@fortawesome/fontawesome-free": "^6.1.1",
-    "@types/lodash": "^4.14.180",
-    "@types/react": "17.0.42",
+    "@types/lodash": "^4.14.181",
+    "@types/react": "17.0.43",
     "@types/react-aria-menubutton": "^6.2.9",
     "@types/react-beautiful-dnd": "^13.1.2",
     "@types/react-dom": "17.0.14",
@@ -44,7 +44,7 @@
     "i18next-scanner": "^3.1.0",
     "react-scripts": "5.0.0",
     "source-map-explorer": "^2.5.2",
-    "typescript": "4.6.2"
+    "typescript": "4.6.3"
   },
   "scripts": {
     "analyze": "source-map-explorer 'build/static/js/*.js'",
